requirements.txt
Django==5.0.6
Django_stdimage==6.0.2
Pillow==10.3.0